+++ src/main/java/org/songeun/petdongne_server/facility/infrastructure/PetFacilityRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.facility.infrastructure;
 
-import org.songeun.petdongne_server.facility.domain.PetFacility;
+import org.songeun.petdongne_server.facility.domain.entity.PetFacility;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
