+++ src/main/java/org/songeun/petdongne_server/global/converter/FacilityTypeConverter.java
@@ -2,7 +2,7 @@ package org.songeun.petdongne_server.global.converter;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
-import org.songeun.petdongne_server.facility.domain.FacilityType;
+import org.songeun.petdongne_server.facility.domain.entity.FacilityType;
 
 @Converter
 public class FacilityTypeConverter implements AttributeConverter<FacilityType, String> {
