+++ src/main/java/org/songeun/petdongne_server/residentialComplex/domain/entity/Area.java
@@ -1,4 +1,4 @@
-package org.songeun.petdongne_server.residentialComplex.domain;
+package org.songeun.petdongne_server.residentialComplex.domain.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
