+++ src/main/java/org/songeun/petdongne_server/residentialComplex/domain/entity/ResidentialComplex.java
@@ -1,4 +1,4 @@
-package org.songeun.petdongne_server.residentialComplex.domain;
+package org.songeun.petdongne_server.residentialComplex.domain.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
