+++ src/main/java/org/songeun/petdongne_server/residentialComplex/infrastructure/AreaRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.residentialComplex.infrastructure;
 
-import org.songeun.petdongne_server.residentialComplex.domain.Area;
+import org.songeun.petdongne_server.residentialComplex.domain.entity.Area;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
