+++ src/main/java/org/songeun/petdongne_server/survey/domain/SurveyOptionStat.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.songeun.petdongne_server.residentialComplex.domain.ResidentialComplex;
+import org.songeun.petdongne_server.residentialComplex.domain.entity.ResidentialComplex;
 import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
